+++ plugin-packs/postcss-preset-env/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changes to PostCSS Preset Env
 
+### 7.7.1 (June 3, 2022)
+
+- Updated `postcss-selector-not` to `6.0.0` (major)
+- Updated `@csstools/postcss-trigonometric-functions` to `1.0.1` (patch)
+- Updated `postcss-attribute-case-insensitive` to `5.0.1` (patch)
+- Updated `postcss-custom-media` to `8.0.1` (patch)
+- Updated `postcss-custom-selectors` to `6.0.2` (patch)
+- Updated `cssdb` to `6.6.3` (patch)
+
 ### 7.7.0 (May 31, 2022)
 
 - Added `@csstools/postcss-trigonometric-functions` <br/> [Check the plugin README](https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-trigonometric-functions#readme) for usage details.